Extract `tab-to-indent` and `one-key-formatting` features
